Improve EnsureAssetsCompiled logic to wait for webpack process to finish
Because there may be situations where a user is running a webpack
process in watch mode in the background that will recompile the
assets, it is possible that if the build time is extremely long,
the user could start tests before the build completes. In this case,
EnsureAssetsCompiled would start compiling the assets again even
though there is already a webpack process doing that.

This commit adds behavior to EnsureAssetsCompiled to check whether
or not the webpack process is running in the background, and will
defer the responsibility of compilation to that process, re-checking
the compiled assets every second until they are up to date.